confd: refactor, relocate container network code to cni.[ch]
